# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface ProviderStatus {
  configured: boolean;
  provider?: string;
}

export interface ReelRequest {
  /** Public Instagram reel/post URL */
  url: string;
}

export type MediaItemType = (typeof MediaItemType)[keyof typeof MediaItemType];

export const MediaItemType = {
  video: "video",
  image: "image",
} as const;

export interface MediaItem {
  type: MediaItemType;
  /** Direct media URL for download */
  url: string;
  thumbnail?: string;
  width?: number;
  height?: number;
  durationSec?: number;
}

export interface ReelResponse {
  shortcode: string;
  username?: string;
  userFullName?: string;
  userAvatar?: string;
  caption?: string;
  likeCount?: number;
  commentCount?: number;
  viewCount?: number;
  takenAt?: string;
  media: MediaItem[];
}

export interface StoriesRequest {
  username: string;
}

export type StoryItemType = (typeof StoryItemType)[keyof typeof StoryItemType];

export const StoryItemType = {
  video: "video",
  image: "image",
} as const;

export interface StoryItem {
  id: string;
  type: StoryItemType;
  url: string;
  thumbnail?: string;
  takenAt?: string;
  durationSec?: number;
}

export interface StoriesResponse {
  username: string;
  userFullName?: string;
  userAvatar?: string;
  stories: StoryItem[];
}

export interface ProfileRequest {
  username: string;
}

export interface ProfileResponse {
  username: string;
  fullName: string;
  biography?: string;
  avatar: string;
  avatarHd?: string;
  isPrivate?: boolean;
  isVerified?: boolean;
  followerCount?: number;
  followingCount?: number;
  postCount?: number;
  externalUrl?: string;
}