# lib/api-zod/src/generated/types/mediaItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaItemType } from "./mediaItemType";

export interface MediaItem {
  type: MediaItemType;
  /** Direct media URL for download */
  url: string;
  thumbnail?: string;
  width?: number;
  height?: number;
  durationSec?: number;
}