# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileResponse {
  username: string;
  fullName: string;
  biography?: string;
  avatar: string;
  avatarHd?: string;
  isPrivate?: boolean;
  isVerified?: boolean;
  followerCount?: number;
  followingCount?: number;
  postCount?: number;
  externalUrl?: string;
}