# lib/api-zod/src/generated/types/providerStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderStatus {
  configured: boolean;
  provider?: string;
}