# lib/api-zod/src/generated/types/reelResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaItem } from "./mediaItem";

export interface ReelResponse {
  shortcode: string;
  username?: string;
  userFullName?: string;
  userAvatar?: string;
  caption?: string;
  likeCount?: number;
  commentCount?: number;
  viewCount?: number;
  takenAt?: string;
  media: MediaItem[];
}