# lib/api-zod/src/generated/types/storiesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { StoryItem } from "./storyItem";

export interface StoriesResponse {
  username: string;
  userFullName?: string;
  userAvatar?: string;
  stories: StoryItem[];
}