# lib/api-zod/src/generated/types/storyItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { StoryItemType } from "./storyItemType";

export interface StoryItem {
  id: string;
  type: StoryItemType;
  url: string;
  thumbnail?: string;
  takenAt?: string;
  durationSec?: number;
}